### 2022-03-19 #4257 @editorialbot
[PRE REVIEW]: SwiftVISA: Controlling Instrumentation with a Swift-based Implementation of the VISA Communication Protocol
**Submitting author:** @ohildreth (<a href="http://orcid.org/0000-0001-5358-9080">Owen Hildreth</a>)
**Repository:** https://github.com/SwiftVISA/SwiftVISASwift.git
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
