### 2022-03-21 #4257 @danielskatz
To check the compilation error, you can click on the error link above.  In general, please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). You can use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment.
