### 2022-03-21 #4257 @danielskatz
The error says "found a tab character that violate indentation while scanning a plain scalar at line 23 column 11 (Psych::SyntaxError)"
