### 2022-03-23 #4257 @danielskatz
`@editorialbot assign me as editor` is also a valid command 🙂
