### 2022-05-27 #4257 @OHildreth
@dhhagan.  I don't know anyone approaching automation this way.  This is a side-project to address some issues I've had controlling my equipment  Most academic-level equipment programming is being done ad-hoc using LabView, which everyone hates.    I know some Swift developers that I've taken workshops with that I can recommend.  They aren't academics but they are amazing programmers.  I also know some people at Los Alamos that are extremely excited about this work because they would love to move away from LabView.  I don't know if they would have github accounts since they aren't programmers.  I've put their contact information below.

Are you looking for code-review or manuscript review?  I've listed a few names that I quickly thought of, but I could probably find more if I knew specifically what you are looking for:

**Mikey Ward**
- Programming Expertise
- Big Nerd Ranch
- https://github.com/wookiee
- ward@bignerdranch.com (I think, I can't find his-email address, he mainly uses Slack and LinkedIn)
- https://www.linkedin.com/in/mikey-ward-75482814/

**Daniel Hooks**
- Group leader at Los Alamos with projects requiring automation
- dhooks@lanl.gov
