### 2022-07-04 #4257 @OHildreth
@dhhagan Thanks for the update.  I was thinking, we could look at the people committing to the pyVISA project: https://github.com/pyvisa/pyvisa/graphs/contributors.  They might have relevant experience for our SwiftVISA project.
