### 2022-07-05 #4257 @MatthieuDartiailh
I can try to give this a look though I am familiar with Swift.

I am not familiar with JOSS review process either. Where can I find the pdf and under which format should submit my review ?
