### 2022-07-16 #4257 @kyleniemeyer
@MatthieuDartiailh JOSS reviews are a bit different than traditional academic journals: the reviews are [checklist driven](https://joss.readthedocs.io/en/latest/review_checklist.html) and take place openly in a GitHub issue like this one (this specific issue is just for getting the review process ready though).

The paper PDF is linked above in a prior comment, copied here for convenience:

👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04257/joss.04257/10.21105.joss.04257.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.04257/joss.04257/10.21105.joss.04257.pdf) 📄 👈

but you can always get the latest version with the command `@editorialbot generate pdf` in a comment.

The software package itself is located at https://github.com/SwiftVISA/SwiftVISASwift.git
