### 2022-07-17 #4257 @MatthieuDartiailh
Currently my main issue with the paper is the claim: `SwiftVISASwift uses CoreSwiftVISA types and protocols and implements communication over TCI/IP instruments`. This claim is overly broad since as far as I can tell this is limited to TCPIP:SOCKET resources and no support is offered for instruments relying on the VXI-11 protocol or the more recent hislip protocol.
