### 2022-07-25 #4257 @OHildreth
@MatthieuDartiailh
I must not have worded things as clearly as I thought in the manuscript. I'm happy to fix that.  The project is divided into multiple Repositories and evolutions.  The original one, SwiftVISA, wraps National Instruments pre-compiled VISA framework and has most of the functionality that pyVISA does (USB, TCI/IP, etc.).

Using the original SwiftVISA requires both installing the VISA framework and recompiling SwiftVISA as you change Swift version.  Additionally, it acts as a Framework, which is more work to use than a Swift Package.  To address the Framework vs. Swift Package issue, we developed the NISwiftVISA along with the NISwiftVISAService.  This still uses the underlying VISA framework, but has an intermediary NISwiftVISAService that handles communicating to the devices.  Because it is still relying on the VISA framework, it handles communicating with all the main VISA instrument types (USB, TCI/IP, etc.).

NISwiftVISA solved the Package issue, but it had some MAJOR limitations starting with macOS Monterey.  Specifically, the VISA framework not longer works under Monterey because Apple effectively blocked kernel extensions.  If there are work arounds, National Instruments hasn't implemented them yet.  That is why we wrote SwiftVISASwift.  This latest iteration is a pure Swift Implementation.  It doesn't rely on the VISA framework and is imported into your project as a Swift Package.  The reason it is limited to TCI/IP is because 1) Apple's new USB driver system was very poorly documented and 2) the DriverKit entitlement was specific to vendors.  Since we weren't Keysight, we couldn't request a DriverKit entitlement to Keysight devices.  For the time, we settled for implementing the TCI/IP protocol.   This hasn't hindered us or our calibrators since almost every instrument we had in the lab (10+) had an ethernet port.  The few that didn't (mass flow controllers, furnaces) also didn't have USB ports either, so we used an Arudino as a communication bridge.

  To address the USB limitation, we are considering using the recently released ORSSerialPort Package (previously a framework).  That is still under debate because ORSSerialPort is limited to macOS and we are trying to make SwiftVISASwift work with iOS and padOS operating systems.

  I haven't seen the HiSLIP protocol before, but it is something we can look into for the future.  Overall, this has been a big project stretching over 4 years for us mainly done by one student, Connor Barnes.  Reimplementing VISA is a big project and we've tried to provide multiple solutions based upon the needs of the user (HiSLIP for example would likely work with SwiftVISA), easy of use (NiSwiftVISA as a package instead of a framework), and preparing for the future (SwiftVISASwift).

 Overall, I was hoping that this publication would help bring some attention to the project and help get more collaborators to expand it (for example, someone that needs HiSLIP might be able to help implement it).  I recently had some people from Los Alamos out and they were extremely excited about the work, we are currently working on writing an NSF REU to train CS students to write instrumentation software during the year at the University and then do a summer internship at a National Lab to help them update their instrumentation control software using some implementation of SwiftVISA (depending on the equipment and requirements).

  I hope this helps and I'm more than happy to expand the manuscript to make things clearer.  The journal guidelines said to keep the manuscript short and let the repository document the code and intent.  So maybe I was too short in my discussions.
