### 2022-07-25 #4257 @MatthieuDartiailh
Sorry if I gave the impression your paper was unclear. It is not so and I did follow the split between the different packages. Since this is only a pre-review and I don't think the actual issue to do the review was opened I just decided to share the one thing that bothered me and did not expand on the overall quality of the paper.

To make my point clearer, the VISA specification (https://www.ivifoundation.org/downloads/Architecture%20Specifications/vpp43_2022-05-19.pdf) makes a distinction between two kind of resource classes that can be used over TCPIP.

- One is the INSTR resource class. This resource class is not specific to the TCPIP bus and rely on either the VXI-11 protocol (most common case) or the hislip protocol. Typically this resource class does not require a port and its address string is typically of the form TCPIP::192.168.0.10::INSTR. The INSTR resource class is expected to support events (even though pyvisa-py does not support them). As far as I understand you do not support this resource class.
- the other is specific to the TCPIP bus and is the SOCKET resource class. The socket resource does not support any of the advanced mode of VISA (events, status byte reading through non text messages, etc) and is easy to use outside VISA since it does not rely on a specific protocol and any socket in any programming language can be used for communication.

What would make your paper more "correct" in a sense is to explicitly say that SwiftVISASwift only support TCPIP::SOCKET resources and that support for TCPIP::INSTR and USB::INSTR may be added in the future/is planned (whatever wording agrees the best with your actual plans).
