### 2022-07-28 #4257 @OHildreth
@MatthieuDartiailh Thanks for the comment.  That is a great point and easy to address in the manuscript with your suggestion.  We currently only support TCIP::SOCKET as you pointed out and that the ::INSTR is planned for the future.  The ::SOCKET was an easy test case to start with.
