+++ paper.md
@@ -59,7 +59,7 @@ The figure below shows how these implementations work together and provide optio
 
 `CoreSwiftVISA` is a low-level package that provides a base, underlying implementation of SwiftVISA, excluding the communication implementation portion (i.e. TCI/IP, USB, etc.).  This includes defining base types and protocols.  `CoreSwiftVISA` isn't directly used, instead, it is used by higher-level packages, such as `SwiftVISASwift` and `NISwiftVISA`.  Breaking the core components out into a separate package makes it simpler to abstract away implementation details.  This can be used to create custom backends for SCPI-compliant instruments or other types of instruments. 
 
-`SwiftVISASwift` uses `CoreSwiftVISA` types and protocols and implements communication over TCI/IP instruments.  This is a pure Swift, native backend that does not require installation of the VISA or NI-VISA frameworks.  `SwiftVISASwift` is currently limited to communicating with TCI/IP instruments because communication on iOS and macOS devices now requires signed drivers.  While it is possible to use older USB drivers, macOS has started restricting kexts and breaking out the implementation into those that did and didn't require signed drivers was considered more future-proof.
+`SwiftVISASwift` uses `CoreSwiftVISA` types and protocols and implements communication over TCI/IP instruments.  This is a pure Swift, native backend that does not require installation of the VISA or NI-VISA frameworks.  `SwiftVISASwift` is currently limited to communicating with TCI/IP instruments using the IPAddress:SOCKET configuration.  It does not yet implement VX11.  While we are working on implementing USB support, progress is slow because communication on iOS and macOS devices now requires signed drivers.  While it is possible to use older USB drivers, macOS has started restricting kexts and breaking out the implementation into those that did and didn't require signed drivers was considered more future-proof.
 
 `NISwiftVISA` allows for communicating over USB instruments.  This implementation does use the `NI-VISA C` backend and requires that users have `NI-VISA` 20.0 or later installed.  Additionally, users must install `NISwiftVISAService`.[@GH_NISwiftVISAService]  The `NISwiftVISAService` is a process that runs in the background that allows the NISwiftVISA framework to communicate with `NI-VISA`'s `C` framework. This service is necessary because NI only distributes pre-compiled binaries of its NI-VISA framework and, at we wrote NISwiftVISA, Swift Packages did not support dependencies on pre-compiled binaries within a package.  While Swift 5.3 introduced dependencies for binary dependencies, they were limited to XCFrameworks.  NI-VISA is not an XCFramework.  To circumvent this requirement, the `NISwiftVISA` uses interprocess communication that calls the `NISwiftVISAService`, which handles all communicates to the instruments.  `NISwiftVISA` works well on macOS systems, but is not compatible with iOS systems.  Additionally, unless the NI-VISA framework is updated to a universal framework, it will run under ARM systems (e.g. Apple Silicon) through Apple's Rosetta 2 dynamic binary translator.  Just as Apple's Rosetta 1 translator was eventually Obsolete and unavailable, we can expect Rosetta 2 to also be phased out eventually.
 
@@ -105,17 +105,71 @@ To read from the instrument, call `read()` on the instrument:
 		// Handle the error
 	}
 
[…]
