### 2022-08-01 #4257 @OHildreth
@MatthieuDartiailh I've updated the manuscript with a section listing known limitations (such as the TCIP:SOCKET you listed above) and future work.  I also provided some examples of the actor branch of SwiftVISASwift that shows how to easily access resources asynchronously.
