+++ README.md
@@ -140,3 +140,35 @@ do {
   // Could not query or decode
 }
 ```
+
+## Contributions and Comments
+This project is a slow-moving "labor-of-love" for our group and intended additional features, such as USB support, are worked on sporatically based upon need, time, and our ability to deal with Apple's poor documentation.  We would love help and please e-mail me (or to reply to the relevant issue or open a new one).
+
+When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owner of this repository before making a change.
+
+### All Code Changes Happen Through Pull Requests
+
+1. Fork the repo and create your branch from `master`.
+2. Make sure the syle of your code is consistent with that of the current one (indentation, etc.).
+3. If you've changed any relevant functionalities, update the documentation.
+4. Ensure the application is working correctly.
+5. Issue that pull request.
+
+### Code of Conduct
+
+Use common sense (source: https://github.com/gasparl/possa/blob/master/CONTRIBUTING.md)
+
+Examples:
+
+* Be respectful of differing viewpoints and experiences
+* Gracefully accept constructive criticism
+* Focus on what is best for the community
+* Have empathy towards other community members
+
+Examples of unacceptable behavior by participants include:
+
+* Trolling, insulting/derogatory comments, and personal or political attacks
+* Public or private harassment
+* Publishing others' private information without explicit permission
+* Other conduct which could reasonably be considered inappropriate in a
+  professional setting
