### 2022-09-10 #4752 @editorialbot
[REVIEW]: SwiftVISA: Controlling Instrumentation with a Swift-based Implementation of the VISA Communication Protocol
**Submitting author:** @ohildreth (<a href="http://orcid.org/0000-0001-5358-9080">Owen Hildreth</a>)
**Repository:** https://github.com/SwiftVISA/SwiftVISASwift.git
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.0
