### 2022-09-15 #4752 @OHildreth
Hi @dhhagan,

  I noticed that some of the repositories within the SwiftVISA organization were missing the MIT license.  I've fixed that and all of the repositories now have a dedicated LICENSE file containing a plain-text MIT license.
