### 2022-09-29 SwiftVISA/SwiftVISASwift#1 @MatthieuDartiailh
JOSS paper: pyvisa reference
The usual syntax for PyVISA is this one not pyVISA and PyVISA-py not pyVISApy. The paper use the right casing for the reference to PyVISA-py but not in other occurrences nor for PyVISA

Also note that both PyVISA and PyVISA-py have documentation hosted on readthedocs but the paper currently use readthedocs for PyVISA and github for PyVISA-py

https://github.com/openjournals/joss-reviews/issues/4752
