### 2022-09-29 SwiftVISA/SwiftVISASwift#3 @MatthieuDartiailh
JOSS paper: VISA description
I am bothered by the strong link made between VISA and SCPI. VISA provides an abstraction over the physical bus used for communication but does not mandate the format of the message circulating on that bus (except for IEEE 488.2 messages). In particular VISA also provide access to register based instruments for which SCPI makes no sense at all.

I am not sure how this should be phrased to be palatable for the reader but I believe that such a strong implicit coupling should be avoided as it tends to create some confusions.

https://github.com/openjournals/joss-reviews/issues/4752
