### 2022-09-29 SwiftVISA/SwiftVISASwift#5 @MatthieuDartiailh
JOSS paper: typos 
concurrancy -> concurrency
line 188 geoemtry -> geometry

https://github.com/openjournals/joss-reviews/issues/4752
