### 2022-09-29 SwiftVISA/SwiftVISASwift#6 @MatthieuDartiailh
JOSS paper: overlong lines
112, 118, 119

https://github.com/openjournals/joss-reviews/issues/4752
