### 2022-09-29 SwiftVISA/SwiftVISASwift#7 @MatthieuDartiailh
JOSS paper: discussion of GPIB support
The "current limitation and future work" section discusses USB, VXI-11 (misspelled in VX11)  but does not address GPIB support. While I assume PCIe card support make little sense, GPIB adapters could be supported through USB-adapters.

https://github.com/openjournals/joss-reviews/issues/4752
