### 2022-09-29 SwiftVISA/SwiftVISASwift#8 @MatthieuDartiailh
JOSS paper: authors contribution
While the repository contain contribution from both the first and last authors I could not find contributions from the other three. Would adding a section/comment detailing the contributions of each authors as is done in other journals be possible ?

https://github.com/openjournals/joss-reviews/issues/4752
