### 2022-11-01 #4752 @dhhagan
HI @jarrah42 - did you have issues creating your checklist? Happy to help if needed. Feel free to ping me here or email at david.hagan@quant-aq.com
