### 2022-11-03 7cbad6d
Update paper.md
Updated manuscript to address reviewer comments
