+++ paper.md
@@ -24,25 +24,38 @@ affiliations:
     index: 1
   - name: Consolidated Analysis Center, Incorporated
     index: 2
-date: 11 November 2021
+date: 20 October 2022
 bibliography: bibliography.bib
 
 ---
 
 # Summary
 
-The Virtual Instrument Software Architecture (VISA)[@NI-VISA_download, @wiki_VISA] is a simple Application Programming Interface (API) implementing the Standard Commands for Programmable Instruments (SCPI)[@SCPI] to control test and measurement instrumentation from a computer.  Over the years, VISA has been wrapped or ported to other languages in an effort to make it easier to write programs to control instruments and analyze their data in real time.  This paper introduces `SwiftVISASwift`,[@GH_SwiftVISASwift] a pure Swift-based VISA implementation for use on both x86 and ARM-based processors.  `SwiftVISASwift` allows scientists and engineers to write full-featured, native iOS and macOS applications while leveraging standardized SCPI commands to control their instrumentations.  Unlike other VISA implementations, such as `pyVISA`, `SwiftVISASwift` provides native access to standard iOS and macOS API's and works well with asynchronous programming methodologies.  This paper details the design and basic use of SwiftVISA.
+The Virtual Instrument Software Architecture (VISA)[@GH_VISA_overview, @NI-VISA_download, @wiki_VISA] is a simple Application Programming Interface (API) to communicate with test and measurement instrumentation from a computer.  VISA includes specifications for communicating with resources or instruments over GPIB (General Purpose Interface Bus, IEEE-488) and VXI (VME eXtensions for instrumentation), which are test and measurement specific I/O interfaces along with providing protocols for communicating over PC-Standard I/O standards, such as VXI-11 (over TCP/IP), UCSBTMC (USBTest and Measurement Class, over USB), HiSLIP (High Speed LAN Instrument Protocol).[@wiki_VISA]  VISA's ability to communicate with a wide variety of instruments over a broad range of I/O's using a common set of APIs makes it an attractive communication API for scientists and equipment manufacturers to write equipment control software with.
+
+Over the years, VISA has been wrapped or ported to other languages in an effort to make it easier to write programs to control instruments and analyze their data in real time.  This paper introduces `SwiftVISASwift`,[@GH_SwiftVISASwift] a pure Swift-based VISA implementation for use on both x86 and ARM-based processors.  `SwiftVISASwift` allows scientists and engineers to write full-featured, native iOS and macOS applications while leveraging standardized SCPI commands to control their instrumentations.  Unlike other VISA implementations, such as `PyVISA`, `SwiftVISASwift` provides native access to standard iOS and macOS APIs and works well with asynchronous programming methodologies.  This paper details the design and basic use of packages and frameworks in the SwiftVISA project.
 
 
 # Statement of Need
 
-The VISA framework is a key backbone for many research instruments.  It's SCPI syntax is versatile, easy to learn, and adaptable to low and high throughput data streams.   It's original C implementation makes it easy to wrap a pre-compiled framework (such as `NI-VISA`) in different languages.  `pyVISA` is one of the most widely known and used implementation of this strategy.[@pyVISA]  `pyVISA` is highly portable between operating systems and python itself is relatively easy to learn.  However, wrapping the VISA framework comes with the limitations associated with any wrapper.  For example, the VISA framework hasn't been compiled for ARM processors and `pyVISA` (and other wrappers) are limited to x86 processors.  Additionally, pre-compiled frameworks aren't as portable as packaged-based distribution methods.  `pyVISApy` has worked to address some of these issues by re-implementing VISA entirely in Python.[@GH_pyVISApy]  However, `pyVISApy` still doesn't have  access to the Operating System's APIs.
[…]
