### 2022-11-03 9ca2cc5
Update bibliography.bib
Updated bibliography to include additional references addressing reviewer comments.
