+++ bibliography.bib
@@ -1,27 +1,77 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Owen Hildreth at 2021-12-18 19:41:09 -0700 
+%% Created for Owen Hildreth at 2022-10-28 22:11:50 -0600 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
 
+@webpage{GH_JuliaWrapper,
+	date-added = {2022-10-28 22:11:29 -0600},
+	date-modified = {2022-10-28 22:11:48 -0600},
+	lastchecked = {10/25/2022},
+	month = {05},
+	url = {https://github.com/PainterQubits/VISA.jl},
+	year = {2019}}
+
+@webpage{GH_GOwrapper,
+	date-added = {2022-10-28 22:10:23 -0600},
+	date-modified = {2022-10-28 22:10:55 -0600},
+	lastchecked = {10/25/2022},
+	month = {10},
+	url = {https://github.com/jpoirier/visa},
+	year = {2018}}
+
+@webpage{GH_VISA_Background,
+	date-added = {2022-10-28 21:41:44 -0600},
+	date-modified = {2022-10-28 21:42:14 -0600},
+	lastchecked = {10/25/2022},
+	month = {05},
+	url = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visaoverviewbackground.html},
+	year = {2022}}
+
+@webpage{GH_VISA_Terminology_Resource,
+	date-added = {2022-10-28 21:25:27 -0600},
+	date-modified = {2022-10-28 21:25:53 -0600},
+	lastchecked = {10/25/2022},
+	month = {05},
+	url = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visa_resource_types.html},
+	year = {2022}}
+
+@webpage{GH_VISA_overview,
+	date-added = {2022-10-28 21:06:56 -0600},
+	date-modified = {2022-10-28 21:07:29 -0600},
+	lastchecked = {10/25/2022},
+	month = {10},
+	url = {https://www.ni.com/en-us/support/documentation/supplemental/06/ni-visa-overview.html},
+	year = {2022}}
+
+@webpage{prologix,
+	author = {Prologix LLC},
+	date-added = {2022-10-25 21:12:01 -0600},
+	date-modified = {2022-10-25 21:12:46 -0600},
+	lastchecked = {10/25/2022},
+	month = {10},
+	url = {http://prologix.biz},
+	year = {2022},
+	bdsk-url-1 = {http://prologix.biz}}
+
 @url{GH_pyVISApy,
 	author = {PyVISA-py},
 	date-added = {2021-12-16 16:39:46 -0700},
-	date-modified = {2021-12-18 19:36:32 -0700},
+	date-modified = {2022-10-25 20:52:39 -0600},
 	keywords = {pyVISA},
-	lastchecked = {12/16/2021},
+	lastchecked = {10/25/2022},
 	title = {PyVISA-py},
-	url = {https://github.com/pyvisa/pyvisa-py},
+	url = {https://pyvisa.readthedocs.io/projects/pyvisa-py/en/latest/},
 	urldate = {2014-01-01},
 	year = {2014},
-	bdsk-url-1 = {https://github.com/pyvisa/pyvisa-py}}
+	bdsk-url-1 = {https://pyvisa.readthedocs.io/projects/pyvisa-py/en/latest/}}
 
 @url{pyVISA,
-	author = {pyVISA},
+	author = {PyVISA},
 	date-added = {2021-12-16 16:39:11 -0700},
 	date-modified = {2021-12-18 19:40:12 -0700},
 	keywords = {pyVISA},
