### 2022-11-03 20f0bf6
Update bibliography.bib
Updated references to be cited properly
