+++ bibliography.bib
@@ -42,7 +42,222 @@
 
 @webpage{GH_VISA_overview,
 	date-added = {2022-10-28 21:06:56 -0600},
-	date-modified = {2022-10-28 21:07:29 -0600},
+	date-modified = {2022-10-28 21:07:29 -0600},%% This BibTeX bibliography file was created using BibDesk.
+%% https://bibdesk.sourceforge.io/
+
+%% Created for Owen Hildreth at 2022-11-03 14:12:50 -0600 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@url{GH_JuliaWrapper,
+	author = {Oskar Painter},
+	date-added = {2022-10-28 22:11:29 -0600},
+	date-modified = {2022-11-03 14:12:43 -0600},
+	lastchecked = {10/25/2022},
+	month = {05},
+	title = {VISA wrapper for Julia},
+	url = {https://github.com/PainterQubits/VISA.jl},
+	urldate = {05/08/2019},
+	year = {2019},
+	bdsk-url-1 = {https://github.com/PainterQubits/VISA.jl}}
+
+@electronic{GH_GOwrapper,
+	author = {Joseph Poirier},
+	date-added = {2022-10-28 22:10:23 -0600},
+	date-modified = {2022-11-03 14:05:11 -0600},
+	lastchecked = {10/25/2022},
+	month = {10},
+	title = {A Go wrapper around National Instruments Virtual Instrument Software Architecture (VISA) drive},
+	url = {https://github.com/jpoirier/visa},
+	urldate = {08/08/2018},
+	year = {2018},
+	bdsk-url-1 = {https://github.com/jpoirier/visa}}
+
+@url{GH_VISA_Background,
+	author = {National Instruments},
+	date-added = {2022-10-28 21:41:44 -0600},
+	date-modified = {2022-11-03 14:10:21 -0600},
+	lastchecked = {10/25/2022},
+	month = {05},
+	title = {NI-VISA: Background},
+	url = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visaoverviewbackground.html},
+	urldate = {05/11/2022},
+	year = {2022},
+	bdsk-url-1 = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visaoverviewbackground.html}}
+
+@url{GH_VISA_Terminology_Resource,
+	author = {National Instruments},
+	date-added = {2022-10-28 21:25:27 -0600},
+	date-modified = {2022-11-03 14:11:26 -0600},
+	lastchecked = {10/25/2022},
+	month = {05},
+	title = {VISA Resource Types},
+	url = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visa_resource_types.html},
+	urldate = {05/11/2022},
+	year = {2022},
+	bdsk-url-1 = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visa_resource_types.html}}
+
+@url{GH_VISA_overview,
+	author = {National Instruments},
+	date-added = {2022-10-28 21:06:56 -0600},
+	date-modified = {2022-11-03 14:10:58 -0600},
+	lastchecked = {10/25/2022},
+	month = {10},
+	title = {NI-VISA Overview},
+	url = {https://www.ni.com/en-us/support/documentation/supplemental/06/ni-visa-overview.html},
+	urldate = {10/21/2022},
+	year = {2022},
[…]
