+++ bibliography.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Owen Hildreth at 2022-10-28 22:11:50 -0600 
+%% Created for Owen Hildreth at 2022-11-03 14:19:14 -0600 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -9,53 +9,9 @@
 
 
 @webpage{GH_JuliaWrapper,
-	date-added = {2022-10-28 22:11:29 -0600},
-	date-modified = {2022-10-28 22:11:48 -0600},
-	lastchecked = {10/25/2022},
-	month = {05},
-	url = {https://github.com/PainterQubits/VISA.jl},
-	year = {2019}}
-
-@webpage{GH_GOwrapper,
-	date-added = {2022-10-28 22:10:23 -0600},
-	date-modified = {2022-10-28 22:10:55 -0600},
-	lastchecked = {10/25/2022},
-	month = {10},
-	url = {https://github.com/jpoirier/visa},
-	year = {2018}}
-
-@webpage{GH_VISA_Background,
-	date-added = {2022-10-28 21:41:44 -0600},
-	date-modified = {2022-10-28 21:42:14 -0600},
-	lastchecked = {10/25/2022},
-	month = {05},
-	url = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visaoverviewbackground.html},
-	year = {2022}}
-
-@webpage{GH_VISA_Terminology_Resource,
-	date-added = {2022-10-28 21:25:27 -0600},
-	date-modified = {2022-10-28 21:25:53 -0600},
-	lastchecked = {10/25/2022},
-	month = {05},
-	url = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visa_resource_types.html},
-	year = {2022}}
-
-@webpage{GH_VISA_overview,
-	date-added = {2022-10-28 21:06:56 -0600},
-	date-modified = {2022-10-28 21:07:29 -0600},%% This BibTeX bibliography file was created using BibDesk.
-%% https://bibdesk.sourceforge.io/
-
-%% Created for Owen Hildreth at 2022-11-03 14:12:50 -0600 
-
-
-%% Saved with string encoding Unicode (UTF-8) 
-
-
-
-@url{GH_JuliaWrapper,
 	author = {Oskar Painter},
 	date-added = {2022-10-28 22:11:29 -0600},
-	date-modified = {2022-11-03 14:12:43 -0600},
+	date-modified = {2022-11-03 14:19:09 -0600},
 	lastchecked = {10/25/2022},
 	month = {05},
 	title = {VISA wrapper for Julia},
@@ -64,10 +20,10 @@
 	year = {2019},
 	bdsk-url-1 = {https://github.com/PainterQubits/VISA.jl}}
 
-@electronic{GH_GOwrapper,
+@webpage{GH_GOwrapper,
 	author = {Joseph Poirier},
 	date-added = {2022-10-28 22:10:23 -0600},
-	date-modified = {2022-11-03 14:05:11 -0600},
+	date-modified = {2022-11-03 14:18:49 -0600},
 	lastchecked = {10/25/2022},
 	month = {10},
 	title = {A Go wrapper around National Instruments Virtual Instrument Software Architecture (VISA) drive},
@@ -76,10 +32,10 @@
 	year = {2018},
 	bdsk-url-1 = {https://github.com/jpoirier/visa}}
 
-@url{GH_VISA_Background,
+@webpage{GH_VISA_Background,
[…]
