+++ bibliography.bib
@@ -1,17 +1,17 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Owen Hildreth at 2022-11-03 14:19:14 -0600 
+%% Created for Owen Hildreth at 2022-11-03 14:24:00 -0600 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
 
-@webpage{GH_JuliaWrapper,
+@url{GH_JuliaWrapper,
 	author = {Oskar Painter},
 	date-added = {2022-10-28 22:11:29 -0600},
-	date-modified = {2022-11-03 14:19:09 -0600},
+	date-modified = {2022-11-03 14:22:07 -0600},
 	lastchecked = {10/25/2022},
 	month = {05},
 	title = {VISA wrapper for Julia},
@@ -20,10 +20,10 @@
 	year = {2019},
 	bdsk-url-1 = {https://github.com/PainterQubits/VISA.jl}}
 
-@webpage{GH_GOwrapper,
+@url{GH_GOwrapper,
 	author = {Joseph Poirier},
 	date-added = {2022-10-28 22:10:23 -0600},
-	date-modified = {2022-11-03 14:18:49 -0600},
+	date-modified = {2022-11-03 14:22:02 -0600},
 	lastchecked = {10/25/2022},
 	month = {10},
 	title = {A Go wrapper around National Instruments Virtual Instrument Software Architecture (VISA) drive},
@@ -32,10 +32,10 @@
 	year = {2018},
 	bdsk-url-1 = {https://github.com/jpoirier/visa}}
 
-@webpage{GH_VISA_Background,
+@url{GH_VISA_Background,
 	author = {National Instruments},
 	date-added = {2022-10-28 21:41:44 -0600},
-	date-modified = {2022-11-03 14:18:34 -0600},
+	date-modified = {2022-11-03 14:23:16 -0600},
 	lastchecked = {10/25/2022},
 	month = {05},
 	title = {NI-VISA: Background},
@@ -44,10 +44,10 @@
 	year = {2022},
 	bdsk-url-1 = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visaoverviewbackground.html}}
 
-@webpage{GH_VISA_Terminology_Resource,
+@url{GH_VISA_Terminology_Resource,
 	author = {National Instruments},
 	date-added = {2022-10-28 21:25:27 -0600},
-	date-modified = {2022-11-03 14:18:25 -0600},
+	date-modified = {2022-11-03 14:23:26 -0600},
 	lastchecked = {10/25/2022},
 	month = {05},
 	title = {VISA Resource Types},
@@ -56,10 +56,10 @@
 	year = {2022},
 	bdsk-url-1 = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visa_resource_types.html}}
 
-@webpage{GH_VISA_overview,
+@url{GH_VISA_overview,
 	author = {National Instruments},
 	date-added = {2022-10-28 21:06:56 -0600},
-	date-modified = {2022-11-03 14:18:29 -0600},
+	date-modified = {2022-11-03 14:23:22 -0600},
 	lastchecked = {10/25/2022},
 	month = {10},
 	title = {NI-VISA Overview},
@@ -78,10 +78,10 @@
 	year = {2022},
 	bdsk-url-1 = {http://prologix.biz}}
 
-@webpage{GH_pyVISApy,
+@url{GH_pyVISApy,
 	author = {PyVISA-py},
[…]
