### 2022-11-03 06f7724
Update paper.md
Fixing citations
