+++ paper.md
@@ -31,7 +31,7 @@ bibliography: bibliography.bib
 
 # Summary
 
-The Virtual Instrument Software Architecture (VISA)[@GH_VISA_overview, @NI-VISA_download, @wiki_VISA] is a simple Application Programming Interface (API) to communicate with test and measurement instrumentation from a computer.  VISA includes specifications for communicating with resources or instruments over GPIB (General Purpose Interface Bus, IEEE-488) and VXI (VME eXtensions for instrumentation), which are test and measurement specific I/O interfaces along with providing protocols for communicating over PC-Standard I/O standards, such as VXI-11 (over TCP/IP), UCSBTMC (USBTest and Measurement Class, over USB), HiSLIP (High Speed LAN Instrument Protocol).[@wiki_VISA]  VISA's ability to communicate with a wide variety of instruments over a broad range of I/O's using a common set of APIs makes it an attractive communication API for scientists and equipment manufacturers to write equipment control software with.
+The Virtual Instrument Software Architecture (VISA)[@GH_VISA_overview; @NI-VISA_download; @wiki_VISA] is a simple Application Programming Interface (API) to communicate with test and measurement instrumentation from a computer.  VISA includes specifications for communicating with resources or instruments over GPIB (General Purpose Interface Bus, IEEE-488) and VXI (VME eXtensions for instrumentation), which are test and measurement specific I/O interfaces along with providing protocols for communicating over PC-Standard I/O standards, such as VXI-11 (over TCP/IP), UCSBTMC (USBTest and Measurement Class, over USB), HiSLIP (High Speed LAN Instrument Protocol).[@wiki_VISA]  VISA's ability to communicate with a wide variety of instruments over a broad range of I/O's using a common set of APIs makes it an attractive communication API for scientists and equipment manufacturers to write equipment control software with.
 
 Over the years, VISA has been wrapped or ported to other languages in an effort to make it easier to write programs to control instruments and analyze their data in real time.  This paper introduces `SwiftVISASwift`,[@GH_SwiftVISASwift] a pure Swift-based VISA implementation for use on both x86 and ARM-based processors.  `SwiftVISASwift` allows scientists and engineers to write full-featured, native iOS and macOS applications while leveraging standardized SCPI commands to control their instrumentations.  Unlike other VISA implementations, such as `PyVISA`, `SwiftVISASwift` provides native access to standard iOS and macOS APIs and works well with asynchronous programming methodologies.  This paper details the design and basic use of packages and frameworks in the SwiftVISA project.
 
