### 2022-11-03 2161567
Update paper.md
Fixing citations.
Changed "," to ";"
