### 2022-11-03 973f07b
Update bibliography.bib
Continuing to fix citations. 
Updated Authors of PyVISA and PyVISA-py readthdocs pages.
