+++ bibliography.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Owen Hildreth at 2022-11-03 14:24:00 -0600 
+%% Created for Owen Hildreth at 2022-11-03 14:38:38 -0600 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -79,9 +79,9 @@
 	bdsk-url-1 = {http://prologix.biz}}
 
 @url{GH_pyVISApy,
-	author = {PyVISA-py},
+	author = {Matthieu Dartiailh},
 	date-added = {2021-12-16 16:39:46 -0700},
-	date-modified = {2022-11-03 14:22:21 -0600},
+	date-modified = {2022-11-03 14:38:38 -0600},
 	keywords = {pyVISA},
 	lastchecked = {10/25/2022},
 	title = {PyVISA-py},
@@ -91,12 +91,12 @@
 	bdsk-url-1 = {https://pyvisa.readthedocs.io/projects/pyvisa-py/en/latest/}}
 
 @url{pyVISA,
-	author = {PyVISA},
+	author = {Matthieu Dartiailh},
 	date-added = {2021-12-16 16:39:11 -0700},
-	date-modified = {2022-11-03 14:23:54 -0600},
+	date-modified = {2022-11-03 14:37:52 -0600},
 	keywords = {pyVISA},
 	lastchecked = {12/16/2021},
-	title = {pyVISA},
+	title = {pyVISA Documetation},
 	url = {https://pyvisa.readthedocs.io/en/latest/},
 	urldate = {2014-01-01},
 	year = {2014},
