### 2022-11-03 bb46800
Update bibliography.bib
Fixed cite keys for pyVISA to PyVISA and GH_pyVISA-py to GH_PyVISA-py
