+++ bibliography.bib
@@ -78,7 +78,7 @@
 	year = {2022},
 	bdsk-url-1 = {http://prologix.biz}}
 
-@url{GH_pyVISApy,
+@url{GH_PyVISApy,
 	author = {Matthieu Dartiailh},
 	date-added = {2021-12-16 16:39:46 -0700},
 	date-modified = {2022-11-03 14:38:38 -0600},
@@ -90,7 +90,7 @@
 	year = {2014},
 	bdsk-url-1 = {https://pyvisa.readthedocs.io/projects/pyvisa-py/en/latest/}}
 
-@url{pyVISA,
+@url{PyVISA,
 	author = {Matthieu Dartiailh},
 	date-added = {2021-12-16 16:39:11 -0700},
 	date-modified = {2022-11-03 14:37:52 -0600},
