+++ paper.md
@@ -51,7 +51,7 @@ This section provides a brief overview of VISA to better understand the purpose 
 It is important to understand some of VISA's Resource terminology.[@GH_VISA_Terminology_Resource]  A Resource is an instrument connected to your system and the Resource provides a complete description of the set of capabilities for the instrument.  There are several types of VISA resources: INSTR, MEMACC, INTFC, BACKPLANE, SERVANT, SOCKET, and RAW.  According to the National Instruments VISA Resources page, most VISA applications and instrument drivers use only the INSTR resource.[@GH_VISA_Terminology_Resource]  The Resource's Interface often describes a concatenation of the communication protocol with the Resource Type. For example, a resource communicating over USB using the Instrument Type would be called a USB::INSTR.  Communicating over TCP/IP using a Socket type would be TCPIP::SOCKET.  
 
 
-While the VISA API provides a simple programming interface to control various T&M instruments, a set of frameworks and drivers are necessary to handle communication between the devices and the computer.  This is often handled using the pre-compiled National Instruments VISA framework, NI-VISA.  This C-based framework can communicate with a broad range of instrument types using a broad range of communication protocols.  Overall, NI-VISA is a good starting place when writing T&M instrument control software. However, writing GUI-based software in C is harder than using more modern programming languages, such as python, C++, Swift, etc.  As a result, there is a large community of programmers and scientists writing wrappers around the NI-VISA framework[@pyVISA; @GH_SwiftVISA; @GH_GOwrapper; @GH_JuliaWrapper] or writing their own VISA implementation[@GH_pyVISApy; @GH_SwiftVISASwift] in their preferred language.
+While the VISA API provides a simple programming interface to control various T&M instruments, a set of frameworks and drivers are necessary to handle communication between the devices and the computer.  This is often handled using the pre-compiled National Instruments VISA framework, NI-VISA.  This C-based framework can communicate with a broad range of instrument types using a broad range of communication protocols.  Overall, NI-VISA is a good starting place when writing T&M instrument control software. However, writing GUI-based software in C is harder than using more modern programming languages, such as python, C++, Swift, etc.  As a result, there is a large community of programmers and scientists writing wrappers around the NI-VISA framework[@PyVISA; @GH_SwiftVISA; @GH_GOwrapper; @GH_JuliaWrapper] or writing their own VISA implementation[@GH_PyVISApy; @GH_SwiftVISASwift] in their preferred language.
 
 
 # SwiftVISASwift Overview
