+++ bibliography.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Owen Hildreth at 2022-11-03 14:38:38 -0600 
+%% Created for Owen Hildreth at 2022-11-03 15:00:43 -0600 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -33,9 +33,9 @@
 	bdsk-url-1 = {https://github.com/jpoirier/visa}}
 
 @url{GH_VISA_Background,
-	author = {National Instruments},
+	author = {{National Instruments}},
 	date-added = {2022-10-28 21:41:44 -0600},
-	date-modified = {2022-11-03 14:23:16 -0600},
+	date-modified = {2022-11-03 14:59:13 -0600},
 	lastchecked = {10/25/2022},
 	month = {05},
 	title = {NI-VISA: Background},
@@ -45,9 +45,9 @@
 	bdsk-url-1 = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visaoverviewbackground.html}}
 
 @url{GH_VISA_Terminology_Resource,
-	author = {National Instruments},
+	author = {{National Instruments}},
 	date-added = {2022-10-28 21:25:27 -0600},
-	date-modified = {2022-11-03 14:23:26 -0600},
+	date-modified = {2022-11-03 14:59:21 -0600},
 	lastchecked = {10/25/2022},
 	month = {05},
 	title = {VISA Resource Types},
@@ -57,9 +57,9 @@
 	bdsk-url-1 = {https://www.ni.com/docs/en-US/bundle/ni-visa/page/ni-visa/visa_resource_types.html}}
 
 @url{GH_VISA_overview,
-	author = {National Instruments},
+	author = {{National Instruments}},
 	date-added = {2022-10-28 21:06:56 -0600},
-	date-modified = {2022-11-03 14:23:22 -0600},
+	date-modified = {2022-11-03 14:59:17 -0600},
 	lastchecked = {10/25/2022},
 	month = {10},
 	title = {NI-VISA Overview},
@@ -69,9 +69,9 @@
 	bdsk-url-1 = {https://www.ni.com/en-us/support/documentation/supplemental/06/ni-visa-overview.html}}
 
 @webpage{prologix,
-	author = {Prologix LLC},
+	author = {{Prologix, LLC.}},
 	date-added = {2022-10-25 21:12:01 -0600},
-	date-modified = {2022-10-25 21:12:46 -0600},
+	date-modified = {2022-11-03 14:59:47 -0600},
 	lastchecked = {10/25/2022},
 	month = {10},
 	url = {http://prologix.biz},
@@ -81,10 +81,10 @@
 @url{GH_PyVISApy,
 	author = {Matthieu Dartiailh},
 	date-added = {2021-12-16 16:39:46 -0700},
-	date-modified = {2022-11-03 14:38:38 -0600},
-	keywords = {pyVISA},
+	date-modified = {2022-11-03 14:50:16 -0600},
+	keywords = {PyVISA},
 	lastchecked = {10/25/2022},
-	title = {PyVISA-py},
+	title = {PyVISA-py Documentation},
 	url = {https://pyvisa.readthedocs.io/projects/pyvisa-py/en/latest/},
 	urldate = {2014-01-01},
 	year = {2014},
@@ -93,19 +93,19 @@
 @url{PyVISA,
 	author = {Matthieu Dartiailh},
 	date-added = {2021-12-16 16:39:11 -0700},
-	date-modified = {2022-11-03 14:37:52 -0600},
-	keywords = {pyVISA},
[…]
