### 2022-11-03 3c82115
Update bibliography.bib
Updated titles for URLs.
