+++ bibliography.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Owen Hildreth at 2022-11-03 15:00:43 -0600 
+%% Created for Owen Hildreth at 2022-11-03 15:07:35 -0600 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -68,23 +68,25 @@
 	year = {2022},
 	bdsk-url-1 = {https://www.ni.com/en-us/support/documentation/supplemental/06/ni-visa-overview.html}}
 
-@webpage{prologix,
+@url{prologix,
 	author = {{Prologix, LLC.}},
 	date-added = {2022-10-25 21:12:01 -0600},
-	date-modified = {2022-11-03 14:59:47 -0600},
+	date-modified = {2022-11-03 15:06:54 -0600},
 	lastchecked = {10/25/2022},
 	month = {10},
+	title = {Manufacturer of GBIP-to-USB and GPIB-to-Ethernet Adapters},
 	url = {http://prologix.biz},
+	urldate = {2022},
 	year = {2022},
 	bdsk-url-1 = {http://prologix.biz}}
 
 @url{GH_PyVISApy,
 	author = {Matthieu Dartiailh},
 	date-added = {2021-12-16 16:39:46 -0700},
-	date-modified = {2022-11-03 14:50:16 -0600},
+	date-modified = {2022-11-03 15:07:35 -0600},
 	keywords = {PyVISA},
 	lastchecked = {10/25/2022},
-	title = {PyVISA-py Documentation},
+	title = {Documentation Page for PyVISA-py},
 	url = {https://pyvisa.readthedocs.io/projects/pyvisa-py/en/latest/},
 	urldate = {2014-01-01},
 	year = {2014},
@@ -93,10 +95,10 @@
 @url{PyVISA,
 	author = {Matthieu Dartiailh},
 	date-added = {2021-12-16 16:39:11 -0700},
-	date-modified = {2022-11-03 14:50:05 -0600},
+	date-modified = {2022-11-03 15:07:15 -0600},
 	keywords = {PyVISA},
 	lastchecked = {12/16/2021},
-	title = {PyVISA Documentation},
+	title = {Documentation Page for PyVISA},
 	url = {https://pyvisa.readthedocs.io/en/latest/},
 	urldate = {2014-01-01},
 	year = {2014},
@@ -129,11 +131,11 @@
 @url{NI-VISA_download,
 	author = {{National Instruments}},
 	date-added = {2021-12-16 16:29:39 -0700},
-	date-modified = {2022-11-03 14:59:26 -0600},
+	date-modified = {2022-11-03 15:05:40 -0600},
 	keywords = {VISA Download},
 	lastchecked = {12/16/2021},
 	month = {07},
-	title = {NI-VISA Download},
+	title = {Download Wepage for NI-VISA},
 	url = {https://www.ni.com/en-us/support/downloads/drivers/download.ni-visa.html#409839},
 	urldate = {2021-12-16},
 	year = {2021},
@@ -142,11 +144,11 @@
 @url{GH_NISwiftVISAService,
 	author = {Connor Barnes},
 	date-added = {2021-12-16 16:28:47 -0700},
-	date-modified = {2022-11-03 14:22:17 -0600},
+	date-modified = {2022-11-03 15:05:01 -0600},
 	keywords = {NISwiftVISAService},
 	lastchecked = {12/16/2021},
 	month = {01},
-	title = {NISwiftVISAService},
+	title = {Source Code and Readme for NISwiftVISAService},
[…]
