### 2022-11-03 SwiftVISA/SwiftVISASwift#8 @OHildreth
We’ve added a section for the author contributions.  The SwiftVISA project started as a Colorado School of Mines’ Computer Science Field Session project.  These students are credited as authors since they demonstrated the that concept works and developed a SwiftVISA wrapper that still works to date.  This includes: Connor Barnes, Luke Henke, Lorena Henke, and Ivan Krukov with Owen Hildreth overseeing the project.  After SwiftVISA was working,  I hired Connor Barnes to write some instrumentation control software using SwiftVISA.  We soon saw the limitations of using a wrapper around the VISA framework and Connor Barnes worked write the rest of the SwiftVISA frameworks, packages, and frameworks.

**Added/Updated Text**

Contributions
Connor Barnes, Luke Henke, Lorena Henke, and Ivan Krukov wrote the original SwiftVISA framework[@GH_SwiftVISA] wrapping the pre-compiled NI-VISA framework[@NI-VISA_download] with Owen Hildreth supervising their work.  Connor Barnes continued to work and develop the rest of the frameworks, services, and Swift packages in the SwiftVISA organization.  Owen Hildreth continues to supervise the work and maintain the repositories.
