### 2022-11-03 SwiftVISA/SwiftVISASwift#7 @OHildreth
**Authors’ Response to Reviewer Comment**
Thanks.  We have fixed the VXI-11 misspelling and updated our text to discuss the use of GPIB adapters.

**Original Text**
There are no plans in `SwiftVISASwift` to directly support `FirewireInstrument` given that this legacy connection is not in widespread use.

**Updated Text**
There are no plans in `SwiftVISASwift` to directly support `GPIBInstrument` or `FirewireInstrument` given that these legacy connections re not in widespread use.  However, we have used GPIB-to-USB and GPIB-to-TCIIP adapters[@prologix] to control our instruments with SwiftVISA` and SwiftVISASwift` respectively.  These adapters were chosen because they don’t require additional drivers, making them platform independent.
