### 2022-11-03 SwiftVISA/SwiftVISASwift#6 @OHildreth
Thanks for pointing this out.  I didn’t notice that code blocks weren’t being properly wrapped.  I went through the document and manually adjusted line widths when necessary.
