### 2022-11-03 df4b504
Update paper.md
Fixed Serial images to stay on the page so they don't wrap.
