### 2022-11-03 SwiftVISA/SwiftVISASwift#5 @OHildreth
Thank you for pointing these out.  I’m sorry I didn’t catch them earlier, I didn’t notice that the Markdown editor I was using didn’t include spell check.  I imported the manuscript into Word and fixed a dozen or so misspellings and typos.  Additionally, I had two other students not familiar with the manuscript or topic review the manuscript for clarity and typos.
