### 2022-11-03 SwiftVISA/SwiftVISASwift#3 @OHildreth
**Authors’ Response to Reviewer Comment**
Thank you, you are correct.  We have removed references to SCPI from the document.

**Original Text**
The Virtual Instrument Software Architecture (VISA)[@NI-VISA_download, @wiki_VISA] is a simple Application Programming Interface (API) ~~implementing the Standard Commands for Programmable Instruments (SCPI)[@SCPI]~~ to control test and measurement instrumentation from a computer.

**Updated Text**
The Virtual Instrument Software Architecture (VISA)[@NI-VISA_download, @wiki_VISA] is a simple Application Programming Interface (API) to control test and measurement instrumentation from a computer.
