### 2022-11-03 SwiftVISA/SwiftVISASwift#1 @OHildreth
Thank you.  We have fixed the spelling on all PyVISA and PyVISA-py.  Additionally, we have updated the citation for PyVISA-py to reference the readthedocs.
