### 2022-11-03 #4752 @OHildreth
@dhhagan
Thanks, it took me some time to fix everything in the manuscript, have some students in my group members review the manuscript for clarity and typos, and get the paper to compile cleanly.

I've responded to all of the comments and closed the simplest ones.  **Should I close the rest of the issues or leave them open** until the Reviewers and Editors have verified that I've sufficiently addressed the comments/issues?
