### 2022-11-06 SwiftVISA/SwiftVISASwift#7 @MatthieuDartiailh
Saying that GPIB is not widely used sounds extremely weird to me given the number of old instruments I met in different labs that do use it and the number of people coming to PyVISA with GPIB specific applications so it may be worth toning down a bit the text.

Regarding adapters it may be worth clarifying what was one and what is envisoned. Prologix adapter are weird since they deviate from the standard behavior of GPIB: you talk to the adapter which then translate for you. It may make sense for you to not want to go further but it may be worth to write explicitly.
