### 2022-11-21 4691661
Update paper.md
Address reviewer comments on VISA overview and lack of GPIB support.
