+++ paper.md
@@ -38,7 +38,7 @@ Over the years, VISA has been wrapped or ported to other languages in an effort 
 
 # Statement of Need
 
-The VISA framework is a key backbone for many research instruments.  Its syntax is versatile, easy to learn, and adaptable to low and high throughput data streams.   It's original C implementation makes it easy to wrap a pre-compiled framework (such as `NI-VISA`) in different languages.  `PyVISA` is a good example of this and is one of the most widely known and used implementations of this strategy.[@PyVISA]  `PyVISA` is highly portable between operating systems and python itself is relatively easy to learn.  However, wrapping the VISA framework comes with the limitations associated with any wrapper.  For example, the VISA framework hasn't been compiled for ARM processors and `PyVISA` (and other wrappers) are limited to x86 processors.  Additionally, pre-compiled frameworks aren't as portable as packaged-based distribution methods.  `PyVISA-py` has worked to address some of these issues by re-implementing VISA entirely in Python.[@GH_PyVISApy]  However, `PyVISA-py` still doesn't have  access to the Operating System's APIs.
+The VISA framework is a key backbone for many research instruments.  Its syntax is versatile, easy to learn, and adaptable to low and high throughput data streams.   It's original C implementation makes it easy to wrap a pre-compiled framework (such as `NI-VISA`) in different languages.  `PyVISA` is a good example of this and is one of the most widely known and used implementations of this strategy.[@PyVISA]  `PyVISA` is highly portable between operating systems and python itself is relatively easy to learn.  However, wrapping the VISA framework comes with the limitations associated with any wrapper.  For example, the VISA framework hasn't been compiled for ARM processors and `PyVISA` (and other wrappers) are limited to x86 processors.  Additionally, pre-compiled frameworks aren't as portable as packaged-based distribution methods.  `PyVISA-py` has worked to address some of these issues by re-implementing VISA entirely in Python.[@GH_PyVISApy]  However, `PyVISA-py` still doesn't have native access to the Operating System's APIs.
 
 Inspired by `PyVISA` and `PyVISA-py`, we have developed a collection of Swift-based wrappers, services, and packages to control VISA-compliant instruments using applications written in Swift for computers running macOS.  
 
@@ -46,18 +46,17 @@ Inspired by `PyVISA` and `PyVISA-py`, we have developed a collection of Swift-ba
 
 # VISA Overview
 
-This section provides a brief overview of VISA to better understand the purpose of this project and some of the key terminology used throughout this publication.[@GH_VISA_Background; @GH_VISA_Terminology_Resource]  VISA was originally standardized through the VXIplug&play alliance and is currently maintained by the IVI Foundation.[@wiki_VISA]  The purpose of the VISA API is to provide a set of standard, consistent architecture to communicate with Test and Measurement (T&M) instruments that is reasonably independent of the interface bus and is consistent across devices of similar types.  For example, a programmer should be able use the same code to control multimeters made by different manufacturers (assuming the  instruments' capabilities are the same).  Additionally, even if the capabilities are different, the API to control similar aspects of different instruments is often the same.  One nice aspect of the API is that many of the VISA Commands use to control instruments consist of simple text strings passed using a standard communication interface (USB, TCP/IP, etc.).
[…]
