+++ bibliography.bib
@@ -1,12 +1,22 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Owen Hildreth at 2022-11-03 15:07:35 -0600 
+%% Created for Owen Hildreth at 2022-11-21 12:02:35 -0700 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@url{GH_RS-VISA,
+	author = {{Rohde & Schwarz USA, Inc.}},
+	date-added = {2022-11-21 12:00:40 -0700},
+	date-modified = {2022-11-21 12:02:29 -0700},
+	lastchecked = {11/21/2022},
+	title = {Rohde & Schwarz VISA Information Page},
+	url = {https://www.rohde-schwarz.com/us/applications/r-s-visa-application-note_56280-148812.html},
+	urldate = {11/21/2022},
+	year = {2022}}
 
 @url{GH_JuliaWrapper,
 	author = {Oskar Painter},
