### 2022-11-21 c9c5f9a
Update paper.md
Fixed references
