+++ paper.md
@@ -46,11 +46,11 @@ Inspired by `PyVISA` and `PyVISA-py`, we have developed a collection of Swift-ba
 
 # VISA Overview
 
-This section provides a brief overview of VISA to better understand the purpose of this project and some of the key terminology used throughout this publication.[@GH_VISA_Background, @GH_VISA_Terminology_Resource]  VISA is a Test and Measurement (T&M) communication protocol originally standardized through the VXIplug&play alliance and is currently maintained by the IVI Foundation.[@wiki_VISA]  This specification is implemented by a number of large T&M equipment manufacturers, including National Instruments (NI), Keysight, Tektronik, and Rohde & Schwarz (R&S) to name a few.  It provides a standard architecture to control instruments that is reasonably independent of manufacturer, interface bus, and is consistent across devices of similar types.  For example, a programmer should be able use the same, or at least similar, code to control multimeters made by different manufacturers (assuming the  instruments' capabilities are the same).  Additionally, even if the capabilities are different, the API to control similar aspects of different instruments is often the same.  
+This section provides a brief overview of VISA to better understand the purpose of this project and some of the key terminology used throughout this publication.[@GH_VISA_Background; @GH_VISA_Terminology_Resource]  VISA is a Test and Measurement (T&M) communication protocol originally standardized through the VXIplug&play alliance and is currently maintained by the IVI Foundation.[@wiki_VISA]  This specification is implemented by a number of large T&M equipment manufacturers, including National Instruments (NI), Keysight, Tektronik, and Rohde & Schwarz (R&S) to name a few.  It provides a standard architecture to control instruments that is reasonably independent of manufacturer, interface bus, and is consistent across devices of similar types.  For example, a programmer should be able use the same, or at least similar, code to control multimeters made by different manufacturers (assuming the  instruments' capabilities are the same).  Additionally, even if the capabilities are different, the API to control similar aspects of different instruments is often the same.  
 
 It is important to understand some of VISA's Resource terminology since `SwiftVISA` and `SwiftVISASwift` don't support all the capabilities specificied in the VISA protocol.[@GH_VISA_Terminology_Resource]  A `Resource` is an instrument connected to your system and the `Resource` provides a complete description of the set of capabilities for the instrument.  There are several types of VISA resources: `INSTR`, `MEMACC`, `INTFC`, `BACKPLANE`, `SERVANT`, `SOCKET`, and `RAW`.  According to the National Instruments VISA Resources page, most VISA applications and instrument drivers use only the INSTR resource.[@GH_VISA_Terminology_Resource]  The Resource's Interface often describes a concatenation of the communication protocol with the Resource Type. For example, a resource communicating over USB using the Instrument Type would be called a `USB::INSTR`.  Communicating over TCP/IP using a Socket type would be `TCPIP::SOCKET`.  `SwiftVISA` currently supports communcation over `USB:INSTR`, `TCPIP::INSTR`, and `TCPIP::SOCKET` while `SwiftVISASwift` is currently limited to communication of `TCPIP::SOCKET`.
 
[…]
