### 2022-11-21 9b2cb98
Update paper.md
Added Future work for Linux platforms.
