### 2022-11-21 #4752 @OHildreth
Hi @dhhagan and @jarrah42 and @MatthieuDartiailh,

  I've updated the text to address @MatthieuDartiailh's comments.  I'm not as familiar with VISA at Matthieu is and I'm using some of the terminology incorrectly.  Hopefully I did an okay job, let me know if there is anything else I can do to improve the manuscript.
