### 2022-11-26 SwiftVISA/SwiftVISASwift#7 @OHildreth
Authors’ Response to Reviewer Comment
Thanks.  We have updated the text to reflect that GPIB is still in use, but that adapters exist

**Original Text**
There are no plans in `SwiftVISASwift` to directly support `GPIBInstrument` or `FirewireInstrument` given that these legacy connections not in widespread use.  However, we have used GPIB-to-USB and GPIB-to-TCIIP adapters[@prologix] to control our instruments with SwiftVISA` and SwiftVISASwift` respectively.  These adapters were chosen because they don’t require additional drivers, making them platform independent.

**Updated Text**
We currently have no plans to update the `SwiftVISA` repositories to directly support `GPIBInstrument` or `FirewireInstrument`; however, contributors are welcome to submit pull-requests to add this functionality.  For our own projects, we use GPIB-to-USB and GPIB-to-TCP/IP adapters[@prologix] by Prologix to control our instruments with `SwiftVISA` and `SwiftVISASwift`.  These adapters work well with `SwiftVISA` and `SwiftVISASwift` because they don’t require installation of additional drivers and because the adapter internally translates the GPIB commands for you.
