### 2022-12-01 SwiftVISA/SwiftVISASwift#7 @MatthieuDartiailh
I am fine with edit.
