### 2022-12-01 #4752 @jarrah42
Hi, since I don't have access to a swift application or server, it's hard to verify the accuracy of the instructions. There is no information on how to run the tests. Can you provide a script so they can be run locally? There is nothing describing how to use the library with any development environment. Can you provide and example using Xcode, or even the command line? An complete example code would be helpful to verify this.

Also, I noticed that some of the code samples in the paper run over the margins so you should fix that.
