### 2022-12-03 63a261c
Update paper.md
Clarifying VISA Overview section.
