### 2022-12-03 79726b1
Update paper.md
Fixed wording.
