+++ paper.md
@@ -50,7 +50,7 @@ This section provides a brief overview of VISA to better understand the purpose 
 
 It is important to understand some of VISA's Resource terminology since `SwiftVISA` and `SwiftVISASwift` don't support all the capabilities specificied in the VISA protocol.[@GH_VISA_Terminology_Resource]  A `Resource` is an instrument connected to your system and the `Resource` provides a complete description of the set of capabilities for the instrument.  There are several types of VISA resources: `INSTR`, `MEMACC`, `INTFC`, `BACKPLANE`, `SERVANT`, and `SOCKET`.  According to the National Instruments VISA Resources page, most VISA applications and instrument drivers use only the INSTR resource.[@GH_VISA_Terminology_Resource]  The Resource's Interface often describes a concatenation of the communication protocol with the Resource Type. For example, a resource communicating over USB using the Instrument Type would be called a `USB::INSTR`.  Communicating over TCP/IP using a Socket type would be `TCPIP::SOCKET`.  `SwiftVISA` currently supports communcation over `USB:INSTR`, `TCPIP::INSTR`, and `TCPIP::SOCKET` while `SwiftVISASwift` is currently limited to communication of `TCPIP::SOCKET`.
 
-VISA is a standard for a library allowing for communication with conforming T&M instruments.  A programmer needs an implementation of the VISA standard to actually communicate between devices and the computer.  The implementation often consists of a library (on macOS this could be a pre-compiled Framework or a Swift Package) to expose a VISA-compliant API for the programmer to use.  It might often include drivers to establish a connection between the computer and the device (for example, macOS requires dedicated drivers to communicate with USB-connected devices, but not TCP/IP-connected devices).  This is often handled using a pre-compiled library and driver set, such as those supplied by the National Instruments NI-VISA Framework[@NI-VISA_download] and Rohde & Schwarz R&S VISA framework[@GH_RS-VISA].  Since the VISA standard is based on C, these manufacturers expose a C interface that can communicate with a wide range of instrument types using a number of communication protocols.  However, writing GUI-based software in C is harder than using more modern programming languages, such as python, C++, Swift, etc.  As a result, there is a large community of programmers and scientists writing wrappers around the NI-VISA framework[@PyVISA; @GH_SwiftVISA; @GH_GOwrapper; @GH_JuliaWrapper] or writing their own VISA implementation[@GH_PyVISApy; @GH_SwiftVISASwift] in their preferred language.  This allows developers to write instrumentation control software using a modern language with access modern GUI and system-level APIs.
[…]
