### 2022-12-03 #4752 @OHildreth
Hi @jarrah42,

  I don't know how you would check the instructions without access to macOS computer, Xcode, and a VISA-compliant instrument.

  A basic set instructions are on the SwiftVISASwift repository: https://github.com/SwiftVISA/SwiftVISASwift.  However, these instructions assume an understanding on how to import a Swift Package dependency into Swift.  To help, I created an example Command Line Utility that connects to a Keysight E36203B power supply and reads its voltage.
 https://github.com/SwiftVISA/SwiftVISASwift/blob/main/ExampleSwiftVISASwift.zip
  You don't need to use this specific model or even a power supply to check the code.   Any TCP/IP, VISA compliant instrument should work, the example prints out the instruments details using the `*IDN?` command, which all VISA compliant equipment responds to.  **You will need to replace the IP address and port with your own instrument's information.**
  NOTE:  I put the connection and reading functionality in functions to make it easier to call throwing functions and print out relevant errors.  This might make the code a little longer, but it is more useful to debug if you are having problems.

 The code is below:
```
import Foundation
import SwiftVISASwift

print("Starting Example")

// You need to set the IPaddress and port to match your instrument's IP address and port
let IPaddress = "169.254.10.1"
let port = 5025

// variables to hold the instrument manager and instrument
var instrumentManager: InstrumentManager?
var instrument: MessageBasedInstrument?

/// Function used  to make an instrument using the supplied IP address and port
func makeInstrument() throws {
    if instrumentManager == nil {
        instrumentManager = InstrumentManager.shared
    }

    instrument = try instrumentManager?.instrumentAt(address: IPaddress, port: port)

}

/// Connect to the instrument by making the instrument and then calling `updateMeasuredValue()` and `getInfo()` to print out the current voltage the power supply is set at and the power supply's information.
func connect() {
    do {
        try makeInstrument()
    } catch  {
        print("Could not connect to instrument at: \(IPaddress) on port: \(port)")
        print(error)
    }

    print("instrument connected: \(String(describing: instrument))")
    updateMeasuredVoltage()
    getInfo()
}

func getInfo() {
    do {
        let localDetails = try instrument?.query("cc") ?? ""
        let details = localDetails.replacingOccurrences(of: ",", with: "\n")
        print("Instrument Details:\n\(details)")
    } catch  {
        print("Could not get instrument information")
        print(error)
    }
}

func updateMeasuredVoltage() {
    do {
        let measuredVoltage = try instrument?.query("SOURCE:VOLTAGE?", as: Double.self)
        print("Measured voltage = \(String(describing: measuredVoltage))")
    } catch {
        print("Could not measure voltage")
        print(error)
    }
}

connect()
```
