### 2022-12-03 #4752 @OHildreth
@MatthieuDartiailh I updated the text responding to the last open comment.  Let me know if this is clearer.

Also, I noticed that your work on the PyVISA-py project.  Have you had any luck getting PyVISA-py to work with macOS 12+?  If so, how did you bypass the kext restrictions for USB devices?
