### 2022-12-04 #4752 @OHildreth
@jarrah42
   In case you can't get a mac and an instrument, I can setup a zoom meeting where I demonstrate the entire process starting from a blank xcode project.  Or I could post a video walk-through.  Would that help or be sufficient?
