### 2022-12-04 #4752 @jarrah42
I have a mac and xcode. I don’t have access to an instrument. Would it be possible to connect two clients together so they exchange a test message?
