### 2022-12-05 #4752 @OHildreth
Hi @jarrah42,
  Unfortunately, I don't know enough about the instrument-side of VISA to know how the instrument is supposed to respond.  Currently, the connection is an inet socket.  I don't know anything about mocking up a inet socket, but you might be able to mock on your system and establish a connection to yourself (or another system).
  If you want, I could setup a zoom meeting and I could run through a setup with you on my computer and one of my instruments.
