### 2022-12-06 #4752 @MatthieuDartiailh
A VISA RAW socket has nothing special to it so you could simply bind a socket to a port and use that as a fake instrument. This is however limited to TCPIP RAW instrument anything else cannot be so easily mocked.
