### 2022-12-06 #4752 @MatthieuDartiailh
I do not have access to a mac so I have zero idea if one can get USB resources to work on Mac. PyVISA-py uses pyusb to handle the USB communication part so if you may get more information by looking up at what they are doing. So far nobody complained to me but I have a feeling the MacOS use base is rather small.
