### 2022-12-06 #4752 @jarrah42
@OHildreth thanks for the code sample. Would you please also provide instructions on building this (either from the command line or using XCode) and running it? I think it would be important to include the sample and instructions in the repository as well. I can write a simple server the receives data from a socket connection and see what I get. Even the initial handshake would be useful I think.
