### 2022-12-06 #4752 @OHildreth
@jarrah42 ,

  All you should have to do is download the project, unzip it, open the project by double-clickign the .xcodeproj file, and then build and run the project using the "play" button.  You will need to update the `IPaddress` and `port` variables as described in the main file comments.  The only error I can envision is a "code signing" error.  The project should be setup to be unsigned and run locally, but you might need to go into the Target → "Signing and Capabilities" and put "Team" to _None_ and "Sign Certificate" to _Sign to run locally_.

  I've recorded a video of the process and just need to edit and compress it for uploading.  That seems easier for a user than trying to describe the Xcode GUI in a text document.  Would that work?
