### 2022-12-06 #4752 @OHildreth
@jarrah42
I've uploaded a video tutorial on setting up SwiftVISASwift from scratch in a new project.  I've also added a link in the README.md on the SwiftVISASwift repository (near the bottom, under the **Example Setup Video** heading) along with the code used to connect and communicate with the instrument.
