### 2022-12-21 6983d10
Expanded installation instructions
Added Installation instructions on adding the SwiftVISASwift package as a dependency using Xcode 14.
