+++ README.md
@@ -22,6 +22,26 @@ dependencies: [
 
 SwiftVISASwift automatically exports [CoreSwiftVISA](https://github.com/SwiftVISA/CoreSwiftVISA), so `import SwiftVISASwift` is sufficient for importing CoreSwiftVISA.
 
+Installation can also be done using Xcode's built-in support for adding Swift Package Dependencies.  An example using Xcode 14.2 is as follows:
+
+First, navigate to the SwiftVISASwift Github page and copy the URL SwiftVISASwift source:
+<img src="https://github.com/SwiftVISA/SwiftVISASwift/blob/main/docs/img/Setup_1_github.png" width="598" height="512">
+
+Second, start a new Xcode project:
+<img src="https://github.com/SwiftVISA/SwiftVISASwift/blob/main/docs/img/Setup_2_createNewProject.png" width="457" height="286">
+
+Third, select the type of application.  SwiftVISASwift has been verified to work on macOS and App, Document App, and Command Line Utilites.  
+<img src="https://github.com/SwiftVISA/SwiftVISASwift/blob/main/docs/img/Setup_3_selectAppType.png" width="378" height="253">
+
+Fourth, naviate to the Project → Package Dependencies and select "+" to begin adding the package dependency.
+<img src="https://github.com/SwiftVISA/SwiftVISASwift/blob/main/docs/img/Setup_4_startPackageDependency.png" width="378" height="253">
+
+Fifth, paste the URL to the SwiftVISASwift source copied earlier from Github:
+<img src="https://github.com/SwiftVISA/SwiftVISASwift/blob/main/docs/img/Setup_5_addSwiftVISASwiftPackage.png" width="378" height="253">
+
+Sixth and last, check your Signings and Capabilities by navigating to Targets → Signings and Capabilities.  The easiest settings to use are #Automatically manage Signing# set to _Yes/Checked_ and #Signing Certificate# to _Sign to Run Locally_.  Other settings can be used if you have an Apple ID Developer Account and Certificate.
+<img src="https://github.com/SwiftVISA/SwiftVISASwift/blob/main/docs/img/Setup_6_signingAndPermissions.png" width="378" height="253">
+
 ## Usage
 
 To create a connection to an instrument over TCPIP, pass the network details to `InstrumentManager.shared.instrumentAt(address:port:)`;
