### 2022-12-21 #4752 @OHildreth
@jarrah42
I've updated the readme.md documentation to include written instructions with screenshots on how to install the SwiftVISASwift package using Xcode 14.2.  Let me know if this helps.
