### 2022-12-22 480790c
Update README.md
Clarified installation instructions.
