### 2022-12-22 cf7b469
Update README.md
Fixed some typos.
