### 2022-12-22 dc36cfc
Update README.md
Updated requirements for the 'actor' branch
