### 2023-02-18 8faca56
Update README.md
Added documentation on how to seek support.
