### 2023-02-18 #4752 @OHildreth
@arfon
I was reading the reviews an noticed that I was missing documentation on how to seek support.  I've updated the read.md with:

### Reporting Issues or Problems
* Please submit an Issue if you have any problems with any SwiftVISA frameworks/packages
* Please submit an Issue if you need any help installing or working with any of the SwiftVISA Frameworks/Packages
