### 2023-03-24 8e0f481
Update paper.md
Fixed spelling mistakes
