### 2023-03-24 #4752 @OHildreth
@dhhagan

- I've double checked the authors and affiliations
- I've made a release of the software (0.2.0-alpha, https://github.com/SwiftVISA/SwiftVISASwift/releases/tag/0.2.0)
- I've archived the release on Zenodo
   -  https://zenodo.org/record/7767610#.ZB3Bgy-B1qs
   - DOI:  10.5281/zenodo.7767610
- I've verified the authors and affiliations and Title on Zenodo match the paper on JOSS
- I've verified that the license is the same (MIT on both Zenodo and on Github)

Thanks for all your help.  Please let me know if there is anythign else I need to do.

Owen
