### 2023-03-29 #4752 @kyleniemeyer
Hi @OHildreth, before we publish the article, can you please clean up the citations in the text? I noticed that many/most need a space separating them from the preceding text, and in other cases they should occur prior to any punctuation (meaning, before a period or comma).
