### 2023-03-29 #4752 @OHildreth
@kyleniemeyer
I fixed the spacing and location of the citations.  I couldn't edit the release build (0.2.0) so I made another release (0.2.1) and all the changes are also on `main`.  Please let me know if that does or doesn't work.
